LayoutTests/svg/custom/getsvgdocument-null.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/js-test.js"></script>
</head>
<body>
    <script>
        description("This tests that 'getSVGDocument' returns null on elements that don't contain SVG documents.");

        var types = ['object', 'iframe', 'embed'];

        types.forEach(function(type) {
            shouldBeNull("document.createElement('" + type + "').getSVGDocument()");
        });
    </script>
</body>
</html>
