third_party/blink/web_tests/platform/generic/svg/custom/getsvgdocument-null-expected.txt

This tests that 'getSVGDocument' returns null on elements that don't contain SVG documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.createElement('object').getSVGDocument() is null
PASS document.createElement('iframe').getSVGDocument() is null
PASS document.createElement('embed').getSVGDocument() is null
PASS successfullyParsed is true

TEST COMPLETE
